api-service/requirements.txt
Flask
requests
azure-cosmos